Классы перенесены в папки. В repository загрузка фильмов из DB, потом из сети. В DB сохраняются два случайных фильма.
