+++ app/build.gradle
@@ -61,4 +61,6 @@ dependencies {
     implementation "androidx.room:room-runtime:$room_version"
     kapt "androidx.room:room-compiler:$room_version"
     implementation "androidx.room:room-ktx:$room_version"
+
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.2.0"
 }
