+++ app/src/main/java/ru/demqn/appname/data/repositories/MoviesRepository.kt
@@ -1,21 +1,32 @@
 package ru.demqn.appname.data.repositories
 
+import androidx.lifecycle.MutableLiveData
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import ru.demqn.appname.data.db.MoviesDAO
+import kotlinx.serialization.ExperimentalSerializationApi
 import ru.demqn.appname.data.Movie
+import ru.demqn.appname.data.db.MoviesDAO
+import ru.demqn.appname.data.network.MoviesNetwork
+import ru.demqn.appname.di.MoviesApplication
+import kotlin.random.Random
 
 class MoviesRepository(private val moviesDAO: MoviesDAO) {
 
-    suspend fun getAllMovies() = withContext(Dispatchers.IO) {
-        moviesDAO.getAllMovies()
-    }
+    val listMoviesRepository = MutableLiveData<List<Movie>>(emptyList())
 
-    suspend fun insert(movie: Movie) {
-        moviesDAO.insert(movie)
-    }
+    @ExperimentalSerializationApi
+    suspend fun getAllMovies() {
 
-    suspend fun deleteAll() {
+        var listMoviesDB: List<Movie>
+        withContext(Dispatchers.IO) { listMoviesDB = moviesDAO.getAllMovies() }
+        listMoviesRepository.value = listMoviesDB
+
+        listMoviesDB = MoviesNetwork().nowPlayingData(MoviesApplication().retrofitMoviesApi)
+        listMoviesRepository.value = listMoviesDB
+
         moviesDAO.deleteALL()
+        moviesDAO.insert(listMoviesDB[Random.nextInt (listMoviesDB.size - 1)])
+        moviesDAO.insert(listMoviesDB[Random.nextInt(listMoviesDB.size-1)])
+
     }
 }
