+++ app/src/main/java/ru/demqn/appname/presentation/presenter/MovieDetailsViewModelFactory.kt
@@ -1,15 +1,15 @@
-package ru.demqn.appname
+package ru.demqn.appname.presentation.presenter
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import ru.demqn.appname.data.network.MoviesNetwork
-import ru.demqn.appname.data.repositories.MoviesRepository
 
-class MoviesListViewModelFactory(private val moviesRepository: MoviesRepository) : ViewModelProvider.Factory {
+class MovieDetailsViewModelFactory :
+    ViewModelProvider.Factory {
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T = when (modelClass) {
-        MoviesListViewModel::class.java -> MoviesListViewModel(MoviesNetwork(), moviesRepository)
+        MovieDetailsViewModel::class.java -> MovieDetailsViewModel(MoviesNetwork())
         else -> throw IllegalArgumentException("$modelClass is not registered ViewModel")
     } as T
 }
