+++ app/src/main/java/ru/demqn/appname/presentation/presenter/MoviesListViewModel.kt
@@ -1,0 +1,21 @@
+package ru.demqn.appname.presentation.presenter
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
+import kotlinx.serialization.ExperimentalSerializationApi
+import ru.demqn.appname.data.Movie
+import ru.demqn.appname.data.repositories.MoviesRepository
+
+class MoviesListViewModel(private val moviesRepository: MoviesRepository) : ViewModel() {
+    private var _movieList = moviesRepository.listMoviesRepository
+    val movieList: LiveData<List<Movie>> get() = _movieList
+
+    @ExperimentalSerializationApi
+    fun getMovies() {
+        viewModelScope.launch {
+            moviesRepository.getAllMovies()
+        }
+    }
+}
